PHP API and completed the function for handleLoginSubmit.
Created an untested PHP API that opens a connection to a database, receives the variables e-mail and password and checks if the user exist. If not an error message is displayed. On the Login component, I've completed the function of acquiring the 2 props (email and password) and sending them to the PHP API. No  Sanitation has been implemented yet.