Added a new component: Sign up Form.
Most of the code is copy and paste from the Login Form and an old website i did last year to help me with the process. Found a small function to auto close sign up form is mouse is clicked outside the form itself. CSS need a focus on once the team is happy with the design.