Password reset card.
Added a password Reset functionality as requested. Created 2 new APIs: first one to check if only e-mail exists then an if statement to send an e-mail to the user personal email address. No testing is done. Will implement testing as soon as the code is pushed. Most of the APi and Password reset functionality are copy and past from other components/API.