New Feature: Signup form with API to insert new users upon registration.
Added Email sanitation and database connection on API sendmailtouser.php
Created REST API that has prepare SQL and Sanitization of Variables given. Checks if user exists, if not a new user is added on the table.
Fixed misspelled database username in usernamechecker.php
Added functionality for the submit button in SignupForm.js
Now all APIs will be located in the API folder.